Add caching (#26)
* Rebase.

* Clean up.

* Updated to use smarter cache dir.
Also documented all the things.

* Remove old comment.

* Added messages during dump to make it clearer what stage it's at.

* Set up tests.
Most things aren't covered if you have the cache (because the early cleanup is handled by the cache). Ideally we should test both with and without a cache, but I want to mock that so we aren't hitting the API a lot... so for now we leave a *lot* of testing to "did the cache process properly?"

* Unexported extra functions and added lookup use.

* Renamed new parameter to "use_lookup".

* Updated man to match new terminology.

* Fixed coverage.

* Removed dev WikipediR dependency.
We still require the overall package for now, but not the version in the hanging PR.

* Styled.

* Added lookup downloader.

* Styler broke a nocov.

* Remove extra space.

* Update .split_ functions to take english_content + word so they don't each hit the API.

* Implemented age-based lookup caching.
Also fixed a few related things. Closes #10. Closes #32. Closes #31.

* Update lookup during use.

* Updated back to logical use_lookup.
Lots of cleanup in this commit.

* 100% coverage!
Except there are nocovs in too many places still but shhhh!

* Fixed long code lines.

* Remove word frequency table.