Bug fixes (#51)
* Deal with alternative spellings more completely.

Closes #36.

* Dealt with contractions.

Also fixed a couple bugs along the way. And dealt with redirects.

* Fix misspellings.

Closes #28.

* Deal with deep links.

Closes #50.

* vnum

* More alt/spelling corner cases.

* More alt/spelling corner cases.

* Correction for multiple alts/misspellings.

Also added a safety check for better exception reporting in .process_word_recursive.

* Better example of multi-alt.

* Slight clean-up based on full process.

This will not change the output, I'm just taking care of the 4 warnings I received this time (out of 1073055 words).

* Fix corner cases.

Things identified by doing the full process.

* Deal with template abuse.

Since we saw it in the wild.